requirements.txt
fastapi[standard]
pydantic
google-cloud-aiplatform[tokenization]
google-generativeai
PyPDF2
